Refactor the project parser
